+++ packages/dive/lib/src/dive_input.dart
@@ -31,9 +31,7 @@ class DiveInput {
 
 class DiveInputs {
   static List<DiveInput> fromType(String typeId) =>
-      DivePluginExt.inputsFromType(typeId);
-  static List<DiveInput> audio() =>
-      obslib.audioInputs().map(DiveInput.fromMap).toList();
-  static List<DiveInput> video() =>
-      obslib.videoInputs().map(DiveInput.fromMap).toList();
+      obslib.inputsFromType(typeId).map(DiveInput.fromMap).toList();
+  static List<DiveInput> audio() => obslib.audioInputs().map(DiveInput.fromMap).toList();
+  static List<DiveInput> video() => obslib.videoInputs().map(DiveInput.fromMap).toList();
 }